move funcs for SQL-interaction to separate script
